Added the code for the satellite transceiver (#2)
This is the example code given to us by Interorbital Systems for transmitting and receiving radio packets using the si4464. This code has the pinout for the satellite and is currently set to transmit the battery voltage. Using this code as well as #3 for the ground station, we successfully sent and received messages.